% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{as.list.QPTIFFImage}
\alias{as.list.QPTIFFImage}
\title{Convert a QPTIFFImage to a named list of 2-D channel matrices}
\usage{
\method{as.list}{QPTIFFImage}(x, ...)
}
\arguments{
\item{x}{A \code{QPTIFFImage}.}

\item{...}{Unused.}
}
\value{
A named list of 2-D matrices, one per channel.
}
\description{
Convert a QPTIFFImage to a named list of 2-D channel matrices
}
