% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{as.QPTIFFImage}
\alias{as.QPTIFFImage}
\alias{as.QPTIFFImage.QPTIFFImage}
\alias{as.QPTIFFImage.array}
\alias{as.QPTIFFImage.DelayedArray}
\title{Coerce a 3-D array to a QPTIFFImage}
\usage{
as.QPTIFFImage(x, ...)

\method{as.QPTIFFImage}{QPTIFFImage}(x, ...)

\method{as.QPTIFFImage}{array}(x, metadata = list(), ...)

\method{as.QPTIFFImage}{DelayedArray}(x, metadata = list(), ...)
}
\arguments{
\item{x}{A 3-D numeric array \code{[H x W x C]}.  For a
\code{QPTIFFImage}, the object is returned unchanged.}

\item{...}{Unused; present for S3 method consistency.}

\item{metadata}{Optional named list of metadata to embed.  Ignored when
\code{x} is already a \code{QPTIFFImage}.}
}
\value{
A \code{QPTIFFImage}.
}
\description{
Attaches the \code{QPTIFFImage} class to any 3-D numeric array
\code{[H x W x C]}.  Channel names are taken from \code{dimnames(x)[[3]]}.
This is the intended way to promote a plain array produced outside
\code{read_qptiff} (e.g., the adjusted output of \code{bgnorm_markers})
into a first-class \code{QPTIFFImage}.
}
\examples{
arr <- array(runif(20 * 20 * 3), dim = c(20, 20, 3),
             dimnames = list(NULL, NULL, c("DAPI", "CD3", "CD8")))
img <- as.QPTIFFImage(arr)
class(img)
}
