% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{.is_lazy_qptiff}
\alias{.is_lazy_qptiff}
\title{True when a QPTIFFImage is backed by a DelayedArray (lazy / on-disk)}
\usage{
.is_lazy_qptiff(x)
}
\description{
True when a QPTIFFImage is backed by a DelayedArray (lazy / on-disk)
}
\keyword{internal}
