% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ome_zarr.R
\name{.is_ome_zarr_store}
\alias{.is_ome_zarr_store}
\title{Does a path point at an OME-Zarr / zarr store?}
\usage{
.is_ome_zarr_store(path)
}
\description{
Does a path point at an OME-Zarr / zarr store?
}
\keyword{internal}
