% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{.new_image_info}
\alias{.new_image_info}
\title{Canonical ImageInfo block (nests camera + scan_resolution)}
\usage{
.new_image_info(
  ...,
  camera = .new_camera_info(),
  scan_resolution = .new_scan_resolution()
)
}
\description{
Canonical ImageInfo block (nests camera + scan_resolution)
}
\keyword{internal}
