% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{.new_qptiff_metadata}
\alias{.new_qptiff_metadata}
\title{Canonical QPTIFFMetadata object}
\usage{
.new_qptiff_metadata(
  slide = .new_slide_info(),
  images = list(),
  acquisition_format = .QPTIFF_UNK,
  raw_xml = ""
)
}
\arguments{
\item{slide}{SlideInfo list.}

\item{images}{List of scenes (see \code{.new_scene}).}

\item{acquisition_format}{One of the format constants.}

\item{raw_xml}{Raw first-page XML / attrs string.}
}
\value{
A \code{"QPTIFFMetadata"} S3 object.
}
\description{
Canonical QPTIFFMetadata object
}
\keyword{internal}
