% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{.new_scale_info}
\alias{.new_scale_info}
\title{Canonical ScaleInfo block (one pyramid level)}
\usage{
.new_scale_info(
  level,
  dims = c("y", "x"),
  downsample_factors = NULL,
  pixel_sizes_um = NULL
)
}
\description{
Canonical ScaleInfo block (one pyramid level)
}
\keyword{internal}
