% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{.new_scan_resolution}
\alias{.new_scan_resolution}
\title{Canonical (empty) ScanResolutionInfo block}
\usage{
.new_scan_resolution(...)
}
\description{
Canonical (empty) ScanResolutionInfo block
}
\keyword{internal}
