% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{.new_scene}
\alias{.new_scene}
\title{Canonical scene (one logical image: FullResolution / Label / Macro / ...)}
\usage{
.new_scene(
  image_info = .new_image_info(),
  channels = list(),
  scales = list(),
  raw_xml = ""
)
}
\description{
Canonical scene (one logical image: FullResolution / Label / Macro / ...)
}
\keyword{internal}
