% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ome.R
\name{.ome_to_metadata}
\alias{.ome_to_metadata}
\title{Build a QPTIFFMetadata object from a parsed OME-XML list}
\usage{
.ome_to_metadata(ome, n_levels = 1L, format = .QPTIFF_OME_TIFF)
}
\description{
Build a QPTIFFMetadata object from a parsed OME-XML list
}
\keyword{internal}
