% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPTIFFArraySeed.R
\name{.read_one_ifd_page_layout}
\alias{.read_one_ifd_page_layout}
\title{Read all pixel-related tags from one IFD}
\usage{
.read_one_ifd_page_layout(f, offset, bigtiff, endian)
}
\description{
Read all pixel-related tags from one IFD
}
\keyword{internal}
