% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{.read_one_ifd_tags}
\alias{.read_one_ifd_tags}
\title{Read tags 270 + 277 from one IFD and return the next-IFD offset}
\usage{
.read_one_ifd_tags(f, offset, bigtiff, endian)
}
\description{
Read tags 270 + 277 from one IFD and return the next-IFD offset
}
\keyword{internal}
