% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPTIFFArraySeed.R
\name{.read_stripped_page}
\alias{.read_stripped_page}
\title{Read a stripped TIFF page (handles uncompressed + DEFLATE)}
\usage{
.read_stripped_page(path, layout, h, w, bps, compression, rows, cols)
}
\description{
Read a stripped TIFF page (handles uncompressed + DEFLATE)
}
\keyword{internal}
