% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPTIFFArraySeed.R
\name{.read_tiled_page}
\alias{.read_tiled_page}
\title{Read a tiled TIFF page (handles uncompressed + DEFLATE)}
\usage{
.read_tiled_page(path, layout, h, w, bps, compression, rows, cols)
}
\description{
Read a tiled TIFF page (handles uncompressed + DEFLATE)
}
\keyword{internal}
