% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ome_zarr.R
\name{.zarr_to_metadata}
\alias{.zarr_to_metadata}
\title{Build a QPTIFFMetadata object from OME-Zarr root attributes}
\usage{
.zarr_to_metadata(attrs, channel_names, colours, pixel_size_um, n_levels)
}
\description{
Build a QPTIFFMetadata object from OME-Zarr root attributes
}
\keyword{internal}
