% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{length.QPTIFFImage}
\alias{length.QPTIFFImage}
\title{Number of channels in a QPTIFFImage}
\usage{
\method{length}{QPTIFFImage}(x)
}
\arguments{
\item{x}{A \code{QPTIFFImage}.}
}
\value{
Integer; the number of channels.
}
\description{
Number of channels in a QPTIFFImage
}
