% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{names<-.QPTIFFImage}
\alias{names<-.QPTIFFImage}
\title{Set the channel names of a QPTIFFImage}
\usage{
\method{names}{QPTIFFImage}(x) <- value
}
\arguments{
\item{x}{A \code{QPTIFFImage}.}

\item{value}{Character vector of length equal to the number of channels.}
}
\value{
\code{x} with updated channel names.
}
\description{
Replaces the third-dimension names (channel names) of a
\code{QPTIFFImage}.  Works for both eager (in-memory) and lazy (on-disk)
objects.
}
\examples{
arr <- array(1:8, dim = c(2, 2, 2),
             dimnames = list(NULL, NULL, c("ch1", "ch2")))
img <- .new_QPTIFFImage(arr)
names(img) <- c("DAPI", "CD3")
names(img)
}
