% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{names.QPTIFFImage}
\alias{names.QPTIFFImage}
\title{Channel names of a QPTIFFImage (third-dimension names of the array)}
\usage{
\method{names}{QPTIFFImage}(x)
}
\description{
Channel names of a QPTIFFImage (third-dimension names of the array)
}
