% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{qpi_channel_names}
\alias{qpi_channel_names}
\title{Channel names of the primary scene}
\usage{
qpi_channel_names(meta)
}
\arguments{
\item{meta}{A \code{QPTIFFMetadata} object.}
}
\value{
Character vector of channel names (possibly empty).
}
\description{
Channel names of the primary scene
}
\examples{
path <- system.file("extdata", "PA_HNC_sample.ome.tiff", package = "bgnormR")
qpi_channel_names(metadata(read_qptiff(path)))
}
