% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qptiff.R
\name{names.QPTIFFImage}
\alias{names.QPTIFFImage}
\title{Channel names of a QPTIFFImage (third-dimension names of the array)}
\usage{
\method{names}{QPTIFFImage}(x)
}
\arguments{
\item{x}{A \code{QPTIFFImage}.}
}
\value{
Character vector of channel names, or \code{NULL}.
}
\description{
Channel names of a QPTIFFImage (third-dimension names of the array)
}
