% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{qpi_is_brightfield}
\alias{qpi_is_brightfield}
\title{Is the primary scene a brightfield (RGB) image?}
\usage{
qpi_is_brightfield(meta)
}
\arguments{
\item{meta}{A \code{QPTIFFMetadata} object.}
}
\value{
Logical scalar.
}
\description{
Is the primary scene a brightfield (RGB) image?
}
\examples{
path <- system.file("extdata", "PA_HNC_sample.qptiff", package = "bgnormR")
qpi_is_brightfield(metadata(read_qptiff(path)))
}
