% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qptiff_metadata.R
\name{qpi_n_levels}
\alias{qpi_n_levels}
\title{Number of pyramid levels of the primary scene}
\usage{
qpi_n_levels(meta)
}
\arguments{
\item{meta}{A \code{QPTIFFMetadata} object.}
}
\value{
Integer scalar (at least 1).
}
\description{
Number of pyramid levels of the primary scene
}
\examples{
path <- system.file("extdata", "PA_HNC_sample.qptiff", package = "bgnormR")
qpi_n_levels(metadata(read_qptiff(path)))
}
